[IMP] Clean code, remove unnecessary method
I have changed some part of the code in order to improve it interpretability. Moreover in order to avoid missconceptions I have removed the _repmat_ method implementation since there is an 'equivalent', there is no need to reimplement Matlab behavour. Finally, I have changed the _crossprod()_ method by %*% operator, despite it can be faster in some calculations; since you use %*% it is preferible to be consistent in all places .